requirements.txt
altgraph==0.17.4
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.8
easyocr==1.7.1
filelock==3.15.1
fonttools==4.53.0
fsspec==2024.6.0
h5py==3.11.0
html5lib==1.1
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
intel-openmp==2021.4.0
Jinja2==3.1.4
keyboard==0.13.5
kiwisolver==1.4.5
langdetect==1.0.9
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
MouseInfo==0.1.3
mplcursors==0.5.3
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
packaging==24.1
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.2.2
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycparser==2.22
PyDirectInput==1.0.4
pydyf==0.10.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pypdfium2==4.30.0
pyperclip==1.8.2
pyphen==0.15.0
PyRect==0.2.0
PyScreeze==0.1.30
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-doctr==0.8.1
pytweening==1.2.0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.9.3
requests==2.32.3
scikit-image==0.23.2
scipy==1.13.1
shapely==2.0.4
six==1.16.0
sympy==1.12.1
tbb==2021.12.0
tifffile==2024.5.22
tinycss2==1.3.0
tomli==2.0.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.1
virtualenv==20.26.2
weasyprint==62.2
webencodings==0.5.1
zipp==3.19.2
zopfli==0.2.3
